# specs/002-rl-trading-agent/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: RL Trading Agent API
  version: 0.1.0
  description: API contracts for operating the RL trading agent on BingX gold perpetual pairs.
servers:
  - url: https://api.example.invalid
security:
  - bearerAuth: []
paths:
...
  /agents/{agentId}:
    get:
      summary: Get agent status
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentStatus"
  /agents/{agentId}/start:
    post:
      summary: Start a trading session
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentStartRequest"
      responses:
        "200":
          description: Agent run started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/pause:
    post:
      summary: Pause trading
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent run paused
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/resume:
    post:
      summary: Resume trading
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent run resumed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/stop:
    post:
      summary: Stop trading
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent run stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/config:
    patch:
      summary: Update agent configuration
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentConfigPatch"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentConfig"
...
  /agents/{agentId}/runs:
    get:
      summary: List agent runs
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentRun"
  /agents/{agentId}/evaluations:
    post:
      summary: Run an evaluation
      parameters:
        - $ref: "#/components/parameters/agentId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluationRequest"
      responses:
        "202":
          description: Evaluation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluationReport"
    get:
      summary: List evaluation reports
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Evaluation reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvaluationReport"
...
  /agents/{agentId}/versions:
    get:
      summary: List agent versions
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Agent versions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AgentVersion"
  /agents/{agentId}/versions/{versionId}/promote:
    post:
      summary: Promote an evaluated model version
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/versionId"
      responses:
        "200":
          description: Version promoted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentVersion"
  /agents/{agentId}/versions/{versionId}/rollback:
    post:
      summary: Roll back to a prior model version
      parameters:
        - $ref: "#/components/parameters/agentId"
        - $ref: "#/components/parameters/versionId"
      responses:
        "200":
          description: Rollback completed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentVersion"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
    versionId:
      name: versionId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    TradingPair:
      type: string
      enum: [Gold-USDT, XAUTUSDT, PAXGUSDT]
...
    AgentStatus:
      type: object
      properties:
        id:
          type: string
        currentRun:
          $ref: "#/components/schemas/AgentRun"
        activeVersion:
          $ref: "#/components/schemas/AgentVersion"
        learningEnabled:
          type: boolean
        killSwitchEnabled:
          type: boolean
        promotionGateStatus:
          type: string
          enum: [pass, fail, unknown]
      required: [id]
    AgentRun:
      type: object
      properties:
        id:
          type: string
        mode:
          type: string
          enum: [paper, live]
        pair:
          $ref: "#/components/schemas/TradingPair"
        status:
          type: string
          enum: [running, paused, stopped]
        startedAt:
          type: string
          format: date-time
        stoppedAt:
          type: string
          format: date-time
        learningEnabled:
          type: boolean
        learningWindowMinutes:
          type: integer
          minimum: 1
        datasetVersionId:
          type: string
          nullable: true
        featureSetVersionId:
          type: string
          nullable: true
      required: [id, mode, pair, status]
    AgentStartRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [paper, live]
        pair:
          $ref: "#/components/schemas/TradingPair"
        riskLimitSetId:
          type: string
        learningEnabled:
          type: boolean
        learningWindowMinutes:
          type: integer
          minimum: 1
        datasetVersionId:
          type: string
        featureSetVersionId:
          type: string
      required: [mode, pair, riskLimitSetId]
    AgentConfig:
      type: object
      properties:
        riskLimitSetId:
          type: string
        learningEnabled:
          type: boolean
        learningWindowMinutes:
          type: integer
        dataSourceConfig:
          $ref: "#/components/schemas/DataSourceConfig"
        killSwitchEnabled:
          type: boolean
        killSwitchReason:
          type: string
          nullable: true
        promotionGateMinWinRate:
          type: number
          nullable: true
        promotionGateMinTrades:
          type: integer
          nullable: true
        promotionGateMaxDrawdown:
          type: number
          nullable: true
        minConfidenceScore:
          type: number
          nullable: true
        allowedSourceIds:
          type: array
          items:
            type: string
          nullable: true
      required: [riskLimitSetId]
    AgentConfigPatch:
      type: object
      properties:
        riskLimitSetId:
          type: string
        learningEnabled:
          type: boolean
        learningWindowMinutes:
          type: integer
        dataSourceConfig:
          $ref: "#/components/schemas/DataSourceConfig"
        killSwitchEnabled:
          type: boolean
        killSwitchReason:
          type: string
        promotionGateMinWinRate:
          type: number
        promotionGateMinTrades:
          type: integer
        promotionGateMaxDrawdown:
          type: number
        minConfidenceScore:
          type: number
        allowedSourceIds:
          type: array
          items:
            type: string
      additionalProperties: false
...
    DataSourceConfig:
      type: object
      properties:
        sources:
          type: array
          items:
            type: object
            properties:
              sourceType:
                type: string
                enum:
                  - bingx_candles
                  - bingx_orderbook
                  - bingx_trades
                  - bingx_funding
                  - bingx_open_interest
                  - bingx_mark_price
                  - bingx_index_price
                  - bingx_ticker
                  - ideas
                  - signals
                  - news
                  - ocr_text
                  - trades
              enabled:
                type: boolean
              freshnessThresholdSeconds:
                type: integer
              minCoveragePct:
                type: number
              maxMissingFields:
                type: integer
              minParseConfidence:
                type: number
            required: [sourceType, enabled]
      required: [sources]
...
    EvaluationRequest:
      type: object
      properties:
        pair:
          $ref: "#/components/schemas/TradingPair"
        periodStart:
          type: string
          format: date-time
        periodEnd:
          type: string
          format: date-time
        agentVersionId:
          type: string
        datasetVersionId:
          type: string
        featureSetVersionId:
          type: string
        decisionThreshold:
          type: number
        windowSize:
          type: integer
        stride:
          type: integer
      required: [pair, periodStart, periodEnd]
...
    EvaluationReport:
      type: object
      properties:
        id:
          type: string
        pair:
          $ref: "#/components/schemas/TradingPair"
        winRate:
          type: number
        netPnlAfterFees:
          type: number
        maxDrawdown:
          type: number
        tradeCount:
          type: integer
        exposureByPair:
          type: object
          additionalProperties:
            type: number
        status:
          type: string
          enum: [pass, fail]
        datasetVersionId:
          type: string
          nullable: true
        datasetHash:
          type: string
          nullable: true
        featureSetVersionId:
          type: string
          nullable: true
        artifactUri:
          type: string
          nullable: true
        backtestRunId:
          type: string
          nullable: true
      required: [id, pair, winRate, netPnlAfterFees, maxDrawdown, tradeCount, status]
    AgentVersion:
      type: object
      properties:
        id:
          type: string
        label:
          type: string
        status:
          type: string
          enum: [draft, evaluating, promoted, retired]
        createdAt:
          type: string
          format: date-time
        artifactUri:
          type: string
        artifactChecksum:
          type: string
        artifactSizeBytes:
          type: integer
        trainingWindowStart:
          type: string
          format: date-time
        trainingWindowEnd:
          type: string
          format: date-time
        algorithmLabel:
          type: string
        hyperparameterSummary:
          type: string
      required: [id, status]
...